Move module.rs files into module/mod.rs
This style puts the root modules closer to the contained modules,
which better aligns the on-disk structure with the module hierarchy.
