--- src/extensions/rgp.rs
@@ -1,4 +1,0 @@
-pub mod report;
-pub mod request;
-
-pub const XMLNS: &str = "urn:ietf:params:xml:ns:rgp-1.0";
